backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
pydantic>=2.6.0
db-sqlite3
google-genai>=0.3.0
websockets>=12.0
python-dotenv>=1.0.0
jinja2>=3.1.0
supabase>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.0
supabase==2.24.0
mangum>=0.17.0
pydantic[email]==2.12.5
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
ruff>=0.3.0
